Extract shared test helpers for MockPopover and createMockContext
Move duplicated MockPopover mock and createMockContext/renderWithContext
helpers into __tests__/testHelpers.tsx, used by 5 test files.
